basm
Enum BasicBlockOp

pub enum BasicBlockOp {

    Push(Operand),
    Pop(Option<DynamicLocation>),
    GetChar(Option<DynamicLocation>),
    PutChar(Operand),
    PutInt(Operand),
    Log(String, Vec<DynamicLocation>),
    Set {
        src: Operand,
        dest: DynamicLocation,
    },
    GetAddr {
        src: DynamicLocation,
        dest: DynamicLocation,
        offset: Option<Operand>,
        negative: bool,
    },
    Add {
        lhs: Operand,
        rhs: Operand,
        dest: DynamicLocation,
    },
    Sub {
        lhs: Operand,
        rhs: Operand,
        dest: DynamicLocation,
    },
    Mul {
        lhs: Operand,
        rhs: Operand,
        dest: DynamicLocation,
    },
    Div {
        lhs: Operand,
        rhs: Operand,
        dest: DynamicLocation,
    },
    Neg {
        src: Operand,
        dest: DynamicLocation,
    },
    Eq {
        lhs: Operand,
        rhs: Operand,
        dest: DynamicLocation,
    },
    Ne {
        lhs: Operand,
        rhs: Operand,
        dest: DynamicLocation,
    },
    Inc(DynamicLocation, Option<u64>),
    Dec(DynamicLocation, Option<u64>),
    HexDump,
    DecimalDump,

}

Variants

Push(Operand)

push R0 gets the value of R0 and pushes it onto the stack push 1 pushes the immediate 1 onto the stack

Pop(Option<DynamicLocation>)

pop R0 pops the top of the stack into R0 pop pops the top of the stack

An immediate may not be used here

GetChar(Option<DynamicLocation>)

Get a character from the input and store it in the operand, if supplied. Otherwise, the character is discarded.

getchar R0 gets a character from the user and stores it in R0. getchar SP[R0] gets a character from the user and stores it in the stack location pointed to by R0.

PutChar(Operand)

Get a character operand and print it.

PutInt(Operand)

Get an integer operand and print it.

Log(String, Vec<DynamicLocation>)

Print a string literal

Set

Set a dynamic location to a value.

Fields

src: Operand

dest: DynamicLocation

GetAddr

Get the effective address of a dynamic location and store it in the destination.

Fields

src: DynamicLocation

dest: DynamicLocation

offset: Option<Operand>

negative: bool

Add

Add two operands and store the result in the destination.

Fields

lhs: Operand

rhs: Operand

dest: DynamicLocation

Sub

Subtract two operands and store the result in the destination.

Fields

lhs: Operand

rhs: Operand

dest: DynamicLocation

Mul

Multiply two operands and store the result in the destination.

Fields

lhs: Operand

rhs: Operand

dest: DynamicLocation

Div

Divide two operands and store the result in the destination.

Fields

lhs: Operand

rhs: Operand

dest: DynamicLocation

Neg

Negate an operand and store the result in the destination.

Fields

src: Operand

dest: DynamicLocation

Eq

Check if two operands are equal and store the result in the destination.

Fields

lhs: Operand

rhs: Operand

dest: DynamicLocation

Ne

Check if two operands are not equal and store the result in the destination.

Fields

lhs: Operand

rhs: Operand

dest: DynamicLocation

Inc(DynamicLocation, Option<u64>)

Increment a dynamic location by an optional amount.

Dec(DynamicLocation, Option<u64>)

Decrement a dynamic location by an optional amount.

HexDump

Print a hexadecimal dump

DecimalDump

Print a decimal dump

Implementations

impl BasicBlockOp

pub fn assemble(&self) -> String